// lib/enum/chinh-tinh.enum.ts
export enum ChinhTinhEnum {
  TuVi = "Tử Vi",
  ThienCo = "Thiên Cơ",
  ThaiDuong = "Thái Dương",
  VuKhuc = "Vũ Khúc",
  ThienDong = "Thiên Đồng",
  LiemTrinh = "Liêm Trinh",
  ThienPhu = "Thiên Phủ",
  ThaiAm = "Thái Âm",
  ThamLang = "Tham Lang",
  CuMon = "Cự Môn",
  ThienTuong = "Thiên Tướng",
  ThienLuong = "Thiên Lương",
  ThatSat = "Thất Sát",
  PhaQuan = "Phá Quân",
  ThienHinh = "Thiên Hình",
  ThienRieu = "Thiên Riêu",
  ThienY = "Thiên Y",
  ThienQuan = "Thiên Quan",
}