// lib/enum/phu-tinh.enum.ts
export enum PhuTinhEnum {
  DiaKhong = "Địa Không",
  DiaKiep = "Địa Kiếp",
  VanXuong = "Văn Xương",
  VanKhuc = "Văn Khúc",
  ThaiTue = "Thái Tuế",
  ThienKhoi = "Thiên Khôi",
  ThienViet = "Thiên Việt",
  LongTri = "Long Trì",
  PhuongCac = "Phượng Các",
  LocTon = "Lộc Tồn",
  KinhDuong = "Kình Dương",
  DaLa = "Đà La",
  HoaLoc = "Hóa Lộc",
  HoaQuyen = "Hóa Quyền",
  HoaKhoa = "Hóa Khoa",
  HoaKy = "Hóa Kỵ",
  ThienKhoc = "Thiên Khốc",
  ThienHu = "Thiên Hư",
  ThienMa = "Thiên Mã",
  HongLoan = "Hồng Loan",
  ThienHy = "Thiên Hỷ",
  TamThai = "Tam Thai",
  BatToa = "Bát Tọa",
  AnQuang = "Ân Quang",
  ThienQuy = "Thiên Quý",
}